adding support for multi-domain screenshot blackout
