+++ packages/driver/package.json
@@ -21,6 +21,7 @@
     "@packages/network": "0.0.0-development",
     "@packages/resolve-dist": "0.0.0-development",
     "@packages/runner": "0.0.0-development",
+    "@packages/runner-shared": "0.0.0-development",
     "@packages/server": "0.0.0-development",
     "@packages/ts": "0.0.0-development",
     "@rollup/plugin-node-resolve": "^13.0.4",
