+++ packages/runner-shared/src/iframe/aut-iframe.js
@@ -394,40 +394,6 @@ export class AutIframe {
     })
   }
 
-  beforeScreenshot = (config) => {
-    // could fail if iframe is cross-origin, so fail gracefully
-    try {
-      if (config.disableTimersAndAnimations) {
-        dom.addCssAnimationDisabler(this._body())
-      }
-
-      _.each(config.blackout, (selector) => {
-        dom.addBlackout(this._body(), selector)
-      })
-    } catch (err) {
-      /* eslint-disable no-console */
-      console.error('Failed to modify app dom:')
-      console.error(err)
-      /* eslint-disable no-console */
-    }
-  }
-
-  afterScreenshot = (config) => {
-    // could fail if iframe is cross-origin, so fail gracefully
-    try {
-      if (config.disableTimersAndAnimations) {
-        dom.removeCssAnimationDisabler(this._body())
-      }
-
-      dom.removeBlackouts(this._body())
-    } catch (err) {
-      /* eslint-disable no-console */
-      console.error('Failed to modify app dom:')
-      console.error(err)
-      /* eslint-disable no-console */
-    }
-  }
-
   startStudio = () => {
     if (studioRecorder.isLoading) {
       studioRecorder.start(this._body()[0])
